IDP-1635 - "No lockout key returned for request" after authentication
https://issues.shibboleth.net/jira/browse/IDP-1635

Also revamped context cleanup option handling.

+++ idp-authn-impl/src/main/java/net/shibboleth/idp/authn/impl/StorageBackedAccountLockoutManager.java
@@ -124,7 +124,7 @@ public class StorageBackedAccountLockoutManager extends AbstractIdentifiableInit
         ComponentSupport.ifInitializedThrowUnmodifiabledComponentException(this);
         
         maxAttemptsLookupStrategy = FunctionSupport.constant(
-                (int) Constraint.isGreaterThan(0, attempts, "Attempts must be greater than zero"));
+                Constraint.isGreaterThan(0, attempts, "Attempts must be greater than zero"));
     }
     
     /**
